move `ACTION_LISTEN_NAME` constant to `shared` package
