+++ rasa/core/policies/fallback.py
@@ -4,7 +4,8 @@ import os
 from typing import Any, List, Text, Optional, Dict, Tuple
 
 from rasa.constants import DOCS_URL_MIGRATION_GUIDE
-from rasa.core.actions.action import ACTION_LISTEN_NAME, ACTION_DEFAULT_FALLBACK_NAME
+from rasa.core.actions.action import ACTION_DEFAULT_FALLBACK_NAME
+from rasa.shared.core.constants import ACTION_LISTEN_NAME
 
 import rasa.utils.io
 from rasa.utils import common as common_utils
