+++ rasa/core/trackers.py
@@ -26,7 +26,7 @@ from rasa.nlu.constants import (
     ENTITY_ATTRIBUTE_GROUP,
 )
 from rasa.shared.core import events
-from rasa.core.actions.action import ACTION_LISTEN_NAME  # pytype: disable=pyi-error
+from rasa.shared.core.constants import ACTION_LISTEN_NAME
 from rasa.core.conversation import Dialogue  # pytype: disable=pyi-error
 from rasa.shared.core.events import (  # pytype: disable=pyi-error
     UserUttered,
