+++ rasa/shared/core/constants.py
@@ -2,3 +2,4 @@ DEFAULT_CATEGORICAL_SLOT_VALUE = "__other__"
 EXTERNAL_MESSAGE_PREFIX = "EXTERNAL: "
 ACTION_NAME_SENDER_ID_CONNECTOR_STR = "__sender_id:"
 IS_EXTERNAL = "is_external"
+ACTION_LISTEN_NAME = "action_listen"
