+++ rasa/shared/core/events.py
@@ -16,6 +16,7 @@ from rasa.shared.core.constants import (
     EXTERNAL_MESSAGE_PREFIX,
     ACTION_NAME_SENDER_ID_CONNECTOR_STR,
     IS_EXTERNAL,
+    ACTION_LISTEN_NAME,
 )
 from rasa.shared.nlu.constants import INTENT_NAME_KEY
 
@@ -548,10 +549,6 @@ class Restarted(Event):
         return self.type_name
 
     def apply_to(self, tracker: "DialogueStateTracker") -> None:
-        from rasa.core.actions.action import (  # pytype: disable=pyi-error
-            ACTION_LISTEN_NAME,
-        )
-
         tracker._reset()
         tracker.trigger_followup_action(ACTION_LISTEN_NAME)
