+++ tests/core/test_ensemble.py
@@ -19,11 +19,8 @@ from rasa.core.trackers import DialogueStateTracker
 from rasa.shared.core.events import UserUttered, ActiveLoop, Event
 
 from tests.core import utilities
-from rasa.core.actions.action import (
-    ACTION_DEFAULT_FALLBACK_NAME,
-    ACTION_RESTART_NAME,
-    ACTION_LISTEN_NAME,
-)
+from rasa.core.actions.action import ACTION_DEFAULT_FALLBACK_NAME, ACTION_RESTART_NAME
+from rasa.shared.core.constants import ACTION_LISTEN_NAME
 from rasa.core.constants import USER_INTENT_RESTART, FORM_POLICY_PRIORITY
 from rasa.shared.core.events import ActionExecuted
 from rasa.core.policies.two_stage_fallback import TwoStageFallbackPolicy
