+++ tests/core/test_trackers.py
@@ -10,7 +10,8 @@ import pytest
 
 import rasa.utils.io
 from rasa.core import training, restore
-from rasa.core.actions.action import ACTION_LISTEN_NAME, ACTION_SESSION_START_NAME
+from rasa.core.actions.action import ACTION_SESSION_START_NAME
+from rasa.shared.core.constants import ACTION_LISTEN_NAME
 from rasa.core.agent import Agent
 from rasa.core.constants import REQUESTED_SLOT
 from rasa.core.domain import Domain
